LPS-75283 Extracting tab items to components
