+++ modules/apps/collaboration/document-library/document-library-web/src/main/java/com/liferay/document/library/web/internal/dynamic/data/mapping/util/DLDDMDisplay.java
@@ -20,21 +20,17 @@ import com.liferay.dynamic.data.mapping.util.BaseDDMDisplay;
 import com.liferay.dynamic.data.mapping.util.DDMDisplay;
 import com.liferay.dynamic.data.mapping.util.DDMDisplayTabItem;
 import com.liferay.portal.kernel.language.LanguageUtil;
-import com.liferay.portal.kernel.portlet.LiferayPortletRequest;
-import com.liferay.portal.kernel.portlet.LiferayPortletResponse;
-import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.PortletKeys;
 
-import java.util.Arrays;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
 import java.util.ResourceBundle;
 
-import javax.portlet.PortletRequest;
-import javax.portlet.PortletURL;
-
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
+import org.osgi.service.component.annotations.ReferencePolicy;
+import org.osgi.service.component.annotations.ReferencePolicyOption;
 
 /**
  * @author Eduardo Garcia
@@ -70,60 +66,7 @@ public class DLDDMDisplay extends BaseDDMDisplay {
 
 	@Override
 	public List<DDMDisplayTabItem> getTabItems() {
-		return Arrays.asList(
-			new DDMDisplayTabItem() {
-
-				@Override
-				public String getTitle(Locale locale) {
-					return LanguageUtil.get(
-						getResourceBundle(locale), "documents-and-media");
-				}
-
-				@Override
-				public String getURL(
-						LiferayPortletRequest liferayPortletRequest,
-						LiferayPortletResponse liferayPortletResponse)
-					throws Exception {
-
-					PortletURL portletURL = _portal.getControlPanelPortletURL(
-						liferayPortletRequest,
-						PortletKeys.DOCUMENT_LIBRARY_ADMIN,
-						PortletRequest.RENDER_PHASE);
-
-					portletURL.setParameter(
-						"mvcRenderCommandName", "/document_library/view");
-
-					return portletURL.toString();
-				}
-
-			},
-			new DDMDisplayTabItem() {
-
-				@Override
-				public String getTitle(Locale locale) {
-					return LanguageUtil.get(
-						getResourceBundle(locale), "document-types");
-				}
-
-				@Override
-				public String getURL(
-						LiferayPortletRequest liferayPortletRequest,
-						LiferayPortletResponse liferayPortletResponse)
-					throws Exception {
-
-					PortletURL portletURL = _portal.getControlPanelPortletURL(
-						liferayPortletRequest,
-						PortletKeys.DOCUMENT_LIBRARY_ADMIN,
-						PortletRequest.RENDER_PHASE);
-
-					portletURL.setParameter(
-						"mvcRenderCommandName",
-						"/document_library/view_file_entry_types");
-
-					return portletURL.toString();
-				}
-
-			});
+		return _ddmDisplayTabItems;
 	}
 
 	@Override
@@ -138,7 +81,12 @@ public class DLDDMDisplay extends BaseDDMDisplay {
 		return true;
 	}
 
-	@Reference
-	private Portal _portal;
+	@Reference(
+		policy = ReferencePolicy.DYNAMIC,
+		policyOption = ReferencePolicyOption.GREEDY,
+		target = "(javax.portlet.name=" + PortletKeys.DOCUMENT_LIBRARY + ")"
+	)
+	private volatile List<DDMDisplayTabItem> _ddmDisplayTabItems =
+		new ArrayList<>();
 
 }
